refactor(node/actuator): impl proposal create skelton, split files
